without try-catch in controller

+++ src/tasks/service/tasks.service.ts
@@ -1,12 +1,12 @@
 import {
   Injectable,
-  BadRequestException,
   NotFoundException,
   InternalServerErrorException,
   Logger,
+  BadRequestException,
 } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
-import { Model, Types } from 'mongoose';
+import { Model } from 'mongoose';
 import { UpdateTaskEmployeeDto } from '../../dto/updateTaskEmployee.dto';
 import { UpdateTaskManagerDto } from '../../dto/updateTaskManager.dto';
 import { CreateTaskDto } from '../../dto/createTask.dto';
@@ -25,11 +25,21 @@ export class TasksService {
 
   public readonly logger = new Logger(TasksService.name);
   async getAllTasks(managerId: string): Promise<Task[]> {
+    if (!managerId) {
+      this.logger.log('managerId was not provided');
+      throw new BadRequestException('managerId is required');
+    }
     return await this.taskModel.find({ managerId }).exec();
   }
   async createTask(task: CreateTaskDto) {
+    if (!task) {
+      this.logger.error('Invalid task data');
+      throw new BadRequestException(' task is required');
+    }
+
     const employees = await Promise.all(
       task.employee.map((id) =>
+
         this.usersService.findOneByUserId(id.toString()),
       ),
     );
@@ -39,8 +49,6 @@ export class TasksService {
     if (nonExistentEmployees.length > 0) {
       this.logger.log('One or more employees not found.');
       throw new NotFoundException('One or more employees not found')
-    } else {
-      this.logger.log('All employees found:', employees);
     }
 
     if (!manager) {
@@ -86,22 +94,22 @@ export class TasksService {
     taskId: string,
     task: UpdateTaskManagerDto | UpdateTaskEmployeeDto,
   ): Promise<Task> {
-    try {
-      const updatedTask = await this.taskModel.findOneAndUpdate(
-        { _id: taskId },
-        task,
-        { new: true },
-      );
-      if (!updatedTask) {
-        throw new NotFoundException(`Task with ID ${taskId} not found`);
-      }
-      return updatedTask;
-    } catch (error) {
-      throw new BadRequestException(error.message);
+    if (!taskId || !task) {
+      this.logger.error('Invalid update task data');
+      throw new BadRequestException('Invalid update task data');
     }
+    const updatedTask = await this.taskModel.findOneAndUpdate({ _id: taskId }, task, { new: true },);
+    if (!updatedTask) {
+      throw new NotFoundException(`Task with ID ${taskId} not found`);
+    }
+    return updatedTask;
+
   }
 
   async deleteTask(taskId: string): Promise<Task> {
+    if (!taskId) {
+      throw new BadRequestException('taskId is required')
+    }
     const deletedTask = await this.taskModel.findOneAndDelete({ _id: taskId });
     if (!deletedTask) {
       throw new NotFoundException(`Task with ID ${taskId} not found`);
@@ -109,11 +117,3 @@ export class TasksService {
     return deletedTask;
   }
 }
-
-
-
-
-
-
-
-
